+++ MatrixGame/src/MatrixRobot.hpp
@@ -111,7 +111,7 @@ public:
 
     void SetOwner(CMatrixMapStatic *ms) { m_Weapon->SetOwner(ms); }
 
-    void CreateEffect(DWORD user, FIRE_END_HANDLER handler, EWeapon type, int cooldown = 0) {
+    void CreateEffect(uintptr_t user, FIRE_END_HANDLER handler, EWeapon type, int cooldown = 0) {
         m_Weapon = (CMatrixEffectWeapon *)CMatrixEffect::CreateWeapon(D3DXVECTOR3(0, 0, 0), D3DXVECTOR3(0, 0, 1), user,
                                                                       handler, type, cooldown);
         if (type == WEAPON_REPAIR) {
