+++ MatrixGame/src/MatrixSkinManager.cpp
@@ -175,7 +175,7 @@ static void obj_shadow(const SSkin *sk) {
     }
 }
 
-static void obj_clear(const SSkin *, [[maybe_unused]] DWORD user_param) {
+static void obj_clear(const SSkin *, [[maybe_unused]] uintptr_t user_param) {
     ASSERT_DX(g_D3DD->SetRenderState(D3DRS_ALPHATESTENABLE, FALSE));
     ASSERT_DX(g_D3DD->SetRenderState(D3DRS_ALPHABLENDENABLE, FALSE));
     ASSERT_DX(g_D3DD->SetRenderState(D3DRS_ZWRITEENABLE, TRUE));
@@ -194,7 +194,7 @@ static void obj_clear(const SSkin *, [[maybe_unused]] DWORD user_param) {
     ASSERT_DX(g_D3DD->SetRenderState(D3DRS_LIGHTING, TRUE));
 }
 
-static void obj_clear3stage(const SSkin *, [[maybe_unused]] DWORD user_param) {
+static void obj_clear3stage(const SSkin *, [[maybe_unused]] uintptr_t user_param) {
     ASSERT_DX(g_D3DD->SetRenderState(D3DRS_ALPHATESTENABLE, FALSE));
     ASSERT_DX(g_D3DD->SetRenderState(D3DRS_ALPHABLENDENABLE, FALSE));
     ASSERT_DX(g_D3DD->SetRenderState(D3DRS_ZWRITEENABLE, TRUE));
@@ -205,7 +205,7 @@ static void obj_clear3stage(const SSkin *, [[maybe_unused]] DWORD user_param) {
     g_D3DD->SetTransform(D3DTS_TEXTURE3, &g_MatrixMap->GetIdentityMatrix());
 }
 
-static void obj_clear5stage(const SSkin *, [[maybe_unused]] DWORD user_param) {
+static void obj_clear5stage(const SSkin *, [[maybe_unused]] uintptr_t user_param) {
     ASSERT_DX(g_D3DD->SetTextureStageState(1, D3DTSS_RESULTARG, D3DTA_CURRENT));
     ASSERT_DX(g_D3DD->SetTextureStageState(1, D3DTSS_TEXCOORDINDEX, 1));
     ASSERT_DX(g_D3DD->SetTextureStageState(2, D3DTSS_TEXCOORDINDEX, 2));
@@ -220,7 +220,7 @@ static void obj_clear5stage(const SSkin *, [[maybe_unused]] DWORD user_param) {
     g_D3DD->SetTransform(D3DTS_TEXTURE5, &g_MatrixMap->GetIdentityMatrix());
 }
 
-static void obj_clear4stage(const SSkin*, [[maybe_unused]] DWORD user_param) {
+static void obj_clear4stage(const SSkin*, [[maybe_unused]] uintptr_t user_param) {
     ASSERT_DX(g_D3DD->SetTextureStageState(1, D3DTSS_RESULTARG, D3DTA_CURRENT));
     ASSERT_DX(g_D3DD->SetTextureStageState(1, D3DTSS_TEXCOORDINDEX, 1));
     ASSERT_DX(g_D3DD->SetTextureStageState(2, D3DTSS_TEXCOORDINDEX, 2));
@@ -234,7 +234,7 @@ static void obj_clear4stage(const SSkin*, [[maybe_unused]] DWORD user_param) {
     g_D3DD->SetTransform(D3DTS_TEXTURE4, &g_MatrixMap->GetIdentityMatrix());
 }
 
-static void obj_clear2stage(const SSkin*, [[maybe_unused]] DWORD user_param) {
+static void obj_clear2stage(const SSkin*, [[maybe_unused]] uintptr_t user_param) {
     ASSERT_DX(g_D3DD->SetRenderState(D3DRS_ALPHATESTENABLE, FALSE));
     ASSERT_DX(g_D3DD->SetRenderState(D3DRS_ALPHABLENDENABLE, FALSE));
     ASSERT_DX(g_D3DD->SetRenderState(D3DRS_ZWRITEENABLE, TRUE));
@@ -248,7 +248,7 @@ static void obj_clear2stage(const SSkin*, [[maybe_unused]] DWORD user_param) {
 
 static void obj_side_tex6(
     const SSkin *sk,
-    DWORD user_param,
+    uintptr_t user_param,
     [[maybe_unused]] int pass)
 {
     SMatrixSkin *ms = (SMatrixSkin *)sk;
@@ -272,7 +272,7 @@ static void obj_side_tex6(
 
 static void obj_side_tex6NA(
     const SSkin *sk,
-    DWORD,
+    uintptr_t,
     [[maybe_unused]] int pass)
 {
     SMatrixSkin *ms = (SMatrixSkin *)sk;
@@ -294,7 +294,7 @@ static void obj_side_tex6NA(
 
 static bool obj_side6(
     const SSkin *sk,
-    [[maybe_unused]] DWORD user_param,
+    [[maybe_unused]] uintptr_t user_param,
     [[maybe_unused]] int pass)
 {
     SMatrixSkin *ms = (SMatrixSkin *)sk;
@@ -360,7 +360,7 @@ static bool obj_side6(
     return false;
 }
 
-static bool obj_side6NA(const SSkin *sk, DWORD, [[maybe_unused]] int pass) {
+static bool obj_side6NA(const SSkin *sk, uintptr_t, [[maybe_unused]] int pass) {
     SMatrixSkin *ms = (SMatrixSkin *)sk;
 
     // gloss present
@@ -417,7 +417,7 @@ static bool obj_side6NA(const SSkin *sk, DWORD, [[maybe_unused]] int pass) {
     return false;
 }
 
-static void obj_side_tex4(const SSkin *sk, DWORD user_param, int pass) {
+static void obj_side_tex4(const SSkin *sk, uintptr_t user_param, int pass) {
     SMatrixSkin *ms = (SMatrixSkin *)sk;
 
     if (pass == 0) {
@@ -466,7 +466,7 @@ static void obj_side_tex4(const SSkin *sk, DWORD user_param, int pass) {
     }
 }
 
-static bool obj_side4(const SSkin *sk, [[maybe_unused]] DWORD user_param, int pass) {
+static bool obj_side4(const SSkin *sk, [[maybe_unused]] uintptr_t user_param, int pass) {
     SMatrixSkin *ms = (SMatrixSkin *)sk;
 
     if (pass == 0) {
@@ -583,7 +583,7 @@ static bool obj_side4(const SSkin *sk, [[maybe_unused]] DWORD user_param, int pa
     return false;
 }
 
-static void obj_side_tex4NA(const SSkin *sk, DWORD, int pass) {
+static void obj_side_tex4NA(const SSkin *sk, uintptr_t, int pass) {
     SMatrixSkin *ms = (SMatrixSkin *)sk;
 
     if (pass == 0) {
@@ -619,7 +619,7 @@ static void obj_side_tex4NA(const SSkin *sk, DWORD, int pass) {
     }
 }
 
-static bool obj_side4NA(const SSkin *sk, DWORD, int pass) {
+static bool obj_side4NA(const SSkin *sk, uintptr_t, int pass) {
     SMatrixSkin *ms = (SMatrixSkin *)sk;
 
     if (pass == 0) {
@@ -729,7 +729,7 @@ static bool obj_side4NA(const SSkin *sk, DWORD, int pass) {
     return false;
 }
 
-static bool obj_side3(const SSkin *sk, [[maybe_unused]] DWORD user_param, int pass) {
+static bool obj_side3(const SSkin *sk, [[maybe_unused]] uintptr_t user_param, int pass) {
     SMatrixSkin *ms = (SMatrixSkin *)sk;
     if (pass == 0) {
         ASSERT_DX(g_D3DD->SetSamplerState(0, D3DSAMP_ADDRESSU, D3DTADDRESS_WRAP));
@@ -814,7 +814,7 @@ static bool obj_side3(const SSkin *sk, [[maybe_unused]] DWORD user_param, int pa
     return false;
 }
 
-static bool obj_side3NA(const SSkin *sk, DWORD, int pass) {
+static bool obj_side3NA(const SSkin *sk, uintptr_t, int pass) {
     SMatrixSkin *ms = (SMatrixSkin *)sk;
     if (pass == 0) {
         ASSERT_DX(g_D3DD->SetSamplerState(0, D3DSAMP_ADDRESSU, D3DTADDRESS_WRAP));
@@ -896,7 +896,7 @@ static bool obj_side3NA(const SSkin *sk, DWORD, int pass) {
     return false;
 }
 
-static void obj_side_tex2(const SSkin *sk, DWORD user_param, int pass) {
+static void obj_side_tex2(const SSkin *sk, uintptr_t user_param, int pass) {
     SMatrixSkin *ms = (SMatrixSkin *)sk;
     if (pass == 0) {
         LPDIRECT3DTEXTURE9 coltex = (LPDIRECT3DTEXTURE9)user_param;
@@ -926,7 +926,7 @@ static void obj_side_tex2(const SSkin *sk, DWORD user_param, int pass) {
     }
 }
 
-static void obj_side_tex2NA(const SSkin *sk, DWORD, int pass) {
+static void obj_side_tex2NA(const SSkin *sk, uintptr_t, int pass) {
     SMatrixSkin *ms = (SMatrixSkin *)sk;
     if (pass == 0) {
         ASSERT_DX(g_D3DD->SetTexture(0, ms->m_Tex->Tex()));
@@ -953,7 +953,7 @@ static void obj_side_tex2NA(const SSkin *sk, DWORD, int pass) {
     }
 }
 
-static bool obj_side2(const SSkin *sk, [[maybe_unused]] DWORD user_param, int pass) {
+static bool obj_side2(const SSkin *sk, [[maybe_unused]] uintptr_t user_param, int pass) {
     SMatrixSkin *ms = (SMatrixSkin *)sk;
     if (pass == 0) {
         ASSERT_DX(g_D3DD->SetSamplerState(0, D3DSAMP_ADDRESSU, D3DTADDRESS_WRAP));
@@ -1035,7 +1035,7 @@ static bool obj_side2(const SSkin *sk, [[maybe_unused]] DWORD user_param, int pa
     return false;
 }
 
-static bool obj_side2NA(const SSkin *sk, DWORD, int pass) {
+static bool obj_side2NA(const SSkin *sk, uintptr_t, int pass) {
     SMatrixSkin *ms = (SMatrixSkin *)sk;
     if (pass == 0) {
         ASSERT_DX(g_D3DD->SetSamplerState(0, D3DSAMP_ADDRESSU, D3DTADDRESS_WRAP));
@@ -1116,7 +1116,7 @@ static bool obj_side2NA(const SSkin *sk, DWORD, int pass) {
     return false;
 }
 
-static void obj_ord_tex4(const SSkin *sk, DWORD user_param, int pass) {
+static void obj_ord_tex4(const SSkin *sk, uintptr_t user_param, int pass) {
     SMatrixSkin *ms = (SMatrixSkin *)sk;
 
     CTextureManaged *basetex;
@@ -1190,7 +1190,7 @@ static void obj_ord_tex4(const SSkin *sk, DWORD user_param, int pass) {
     }
 }
 
-static bool obj_ord4(const SSkin *sk, DWORD user_param, int pass) {
+static bool obj_ord4(const SSkin *sk, uintptr_t user_param, int pass) {
     SMatrixSkin *ms = (SMatrixSkin *)sk;
     if (user_param) {
         CMatrixMapObject *o = (CMatrixMapObject *)user_param;
@@ -1352,7 +1352,7 @@ static bool obj_ord4(const SSkin *sk, DWORD user_param, int pass) {
     return false;
 }
 
-static void obj_ord_tex2(const SSkin *sk, DWORD user_param, int pass) {
+static void obj_ord_tex2(const SSkin *sk, uintptr_t user_param, int pass) {
     SMatrixSkin *ms = (SMatrixSkin *)sk;
     CTextureManaged *basetex;
     if (user_param) {
@@ -1412,7 +1412,7 @@ static void obj_ord_tex2(const SSkin *sk, DWORD user_param, int pass) {
     }
 }
 
-static bool obj_ord2(const SSkin *sk, DWORD user_param, int pass) {
+static bool obj_ord2(const SSkin *sk, uintptr_t user_param, int pass) {
     SMatrixSkin *ms = (SMatrixSkin *)sk;
     if (user_param) {
         CMatrixMapObject *o = (CMatrixMapObject *)user_param;
